requirements.txt
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==3.0.1  # Upgrade to a newer version
transformers==4.41.2
huggingface_hub==0.23.0
torch==2.7.1
scipy==1.14.1
numpy==2.1.2
gunicorn==21.2.0
python-dotenv==1.0.1
crawl4ai==0.6.3
nest_asyncio==1.5.6

faiss-cpu==1.10.0
playwright==1.46.0


requests==2.32.3
pymongo==4.8.0
boto3==1.35.7
elasticsearch==8.15.1
pika==1.3.2
pypdf==4.2.0
python-docx==1.1.2
textract==1.6.5
